counsel projectile
Fix `counsel-projectile-ag' by overriding it completely.

I noticed that the way it handles projectile's ignore lists
doesn't work when you call `shell-quote-argument` on the patterns
to ignore.

This might be a breaking "fix", so I left the original override
there and commented it out.

This also configures `so-long` again.

This adds `crux-recentf-find-file` to the `ivy-rich` transformer list.