hydras: Offsets.
Add an optional param `padding` to `wal/ascii-whale-hydra-offset`
so instead of adding whitespace a number can be passed to align lines.

Remove `golden-ratio` again from `wal/lieutenant-sink` and add it
to `wal/resize-window`.

All optional keys are no hinted and no longer part of the string.

This also binds `crux-find-user-custom-file` to "c" in `wal/general-sink`.