Key Bindings: User key prefixing
The hyper key is no longer a requirement.

A new variable `wal/use-hyper-prefix` can be set to nil before loading
the config to instead use key sequence `C-c w`. The template was
amended accordingly.

This extends to both leader keys and all other bindings.

To support this, a `use-package` normalizer and handler based on the
`:bind` functions was added.

The new keyword is `:wal-bind`.

The change is also done in preparation of better terminal support.