:wal-bind, other functions
Fix `:wal-bind` to autoload functions. Replace copy of handler with an
alias.

Use `seq-filter` for fonts instead of deleting nil after `mapcar`.

Add function `wal/avy-goto-word-in-line` and move
`wal/avy-goto-line-and-char` to legacy file.

Simplify `flyspell` usage.

Wrap directory creation in a function so customization would actually
matter.

Out sub packages in a const and loop over them.

Consider it a user error if an attempt is made to delete a local
variable when there is none.

Make sure running in daemon-mode does not preclude using hyper
bindings.

Opening the readme turns on `auto-fill-mode`.