Fix: Previews using consult and perspective
Fixes live `consult` previews when using `perspective`.

Instead of setting private `consult--buffer-display` to the persp
switch buffer function, we customize action of
`consult--source-buffer` to a new function `wal/consult-persp-switch`.

The perspective source is now appended. Configuring `perspective` is
now done in a function. The package is now also demanded if enabled.