Improvement: Buffers in and outside of perspective
Fixes buffers getting added to perspectives when switching buffers
with consult because of live previews.

The perspective source is default and previews, when mode is active,
the normal one no longer is and doesn't. The normal one also filters
out any perspective buffer.

Moves the perspective source to perspective block.