Configuration: Custom flags
Re-add the custom flag checker from legacy file and use it to check
for flag "--doctor" to make package `use-package` collect statistics
and report on them.

In order to make this possible, package `wal-func` is loaded before
package `wal-key-bindings` and its key bindings are moved to the
latter package.

This also removes some unneeded `:if` keywords from minibuffer
completion packages.

This also changes the bindign for `consult-bookmark` to "u" in
bookmark command map.