feat: language packs
Adds new function `wal/install-language-pack` that installs a pack of
packages associated with a language. These are defined in alist
`wal/language-packs`.

The helper function `wal/install-packages` was adapted to use key
parameters and add new parameter `on-done` that is a string that can
be passed to be printed as a message on successful completion.
Deleting windows is now optional.

This also sets `lsp-dart` to `:ensure nil` as it's language-specific.