improve(dart-mode,lsp-dart): no guides, upper-case r for restart
Adds new custom variables `wal/no-indent-guides` and adds `dart-mode`
to it. `highlight-indent-guides-mode` now calls
`wal/maybe-highlight-indent-guides` instead and checks if the major
mode is not blacklisted.

This also binds R again for dart as that's the default binding for a
full restart normally to so an exception is fine.

Removes ejs files from `web-mode'.