feat(wal-func): macro for conditional universal argument calls
Adds new macro `wal/univ` to create a function that calls one of two
functions depending on whether the `universal-argument` was set.

Also adds new function `wal/univ-p` that just checks for the argument.
This function is used in various places now.

Packages `magit` and `embark` use the new macro to do one of two
things.