improve(corfu,corfu-doc,lsp-ui): load docs if not slow
Adds `dart-mode` to the slow LSP servers again.

The logic in `wal/lsp` was changed to not enable `lps-ui-doc` for
`company`. For `corfu`, `corfu-doc` is only enabled if the major mode
is not considered slow.

Adds new function `wal/slow-lsp-p` to check if the major mode is part
of `wal/lsp-slow-modes`.