package(gdscript-mode): add and configure
This will also register a new project type for `projectile`.

This also adds a workaround from the official README to avoid unknown
notification errors of the LSP server.

Macro `wal/lang-hook` was amended to allow forcing `wal/enable-tabs`
when passing `'always` to the `:tabs` keyword.

This also binds `server-edit` for convenience since I use it in Godot.