feat(windows): tab buffers
Adds the concept of tab buffers. Finding a file will now associate its
buffer with a tab if it has an explicit name. This is stored in a
hash-map where the key is tab's explicit name and the value is a ring
of size 10.

Killing a buffer will remove it from history. Killing a tab will
remove its history.

This is used for a new consult buffer source that can be narrowed to
using t.